Using same code for Windows & Ubuntu

+++ src/openpose/core/nmsBase.cu
@@ -7,8 +7,8 @@
 
 namespace op
 {
-    const auto THREADS_PER_BLOCK_1D = 16;
-    const auto THREADS_PER_BLOCK = 512;
+    const auto THREADS_PER_BLOCK_1D = 16u;
+    const auto THREADS_PER_BLOCK = 512u;
 
     template <typename T>
     __global__ void nmsRegisterKernel(int* kernelPtr, const T* const sourcePtr, const int w, const int h, const T threshold)
@@ -121,9 +121,9 @@ namespace op
             const auto imageOffset = height * width;
             const auto offsetTarget = (maxPeaks+1)*targetSize[3];
 
-            const dim3 threadsPerBlock2D{(unsigned int)THREADS_PER_BLOCK_1D, (unsigned int)THREADS_PER_BLOCK_1D};
+            const dim3 threadsPerBlock2D{THREADS_PER_BLOCK_1D, THREADS_PER_BLOCK_1D};
             const dim3 numBlocks2D{getNumberCudaBlocks(width, threadsPerBlock2D.x), getNumberCudaBlocks(height, threadsPerBlock2D.y)};
-            const dim3 threadsPerBlock1D{ (unsigned int)THREADS_PER_BLOCK};
+            const dim3 threadsPerBlock1D{THREADS_PER_BLOCK};
             const dim3 numBlocks1D{getNumberCudaBlocks(imageOffset, threadsPerBlock1D.x)};
             // log("num_b: " + std::to_string(bottom->shape(0)));       // = 1
             // log("channel_b: " + std::to_string(bottom->shape(1)));   // = 57 = 18 body parts + bkg + 19x2 PAFs
